// src/app/web/api/division.ts
export interface Division {
    divId: string;

    divParentId: string;

    divParentName: string;

    orgId: string;

    orgName: string;

    locId: string;

    locName:string;

    divisionCode: string;

    divisionName: string;

    divisionStatus: string;

    divisionType: string;

    partnerFlag: string;

    createdDate: string;

    createdBy: string;

    updatedDate: string;

    updatedBy: string;

    name: string,

    id: string
}